app/requirements.txt
strawberry-graphql[fastapi]==0.158.0
email-validator==1.3.0
python-dotenv==0.21.0
phonenumbers==8.13.19
python-jose==3.3.0
fastapi==0.100.0
requests==2.28.1
uvicorn==0.18.3
pymongo==4.3.2
passlib==1.7.4
pytest==7.4.0
jinja2==3.1.2
pyotp==2.8.0